refactor(service): split resource_service.go into update + tags
Extract update-payload application into resource_update.go and tag management into resource_tags.go. Same package, same receiver, no signature changes. Closes the ~400-line target feature 037 left unmet (749 -> 409 lines).
