+++ .specify/feature.json
@@ -1,3 +1,3 @@
 {
-  "feature_directory": "specs/037-service-decomposition"
+  "feature_directory": "specs/038-resource-service-split"
 }
