+++ .specify/integrations/claude.manifest.json
@@ -1,7 +1,7 @@
 {
   "integration": "claude",
   "version": "0.7.4.dev0",
-  "installed_at": "2026-05-15T16:32:18.197504+00:00",
+  "installed_at": "2026-05-28T00:31:58.244657+00:00",
   "files": {
     ".claude/skills/speckit-analyze/SKILL.md": "f7b3b921069869fafe0a1f44c17037a74bd11710d030a1961a2995f9f5672dec",
     ".claude/skills/speckit-checklist/SKILL.md": "36f82b9b727bc4a0a60a435b218a6d33c1bb421c8bc05b283cdcfd36b4cf2fb3",
