+++ .specify/integrations/speckit.manifest.json
@@ -1,6 +1,6 @@
 {
   "integration": "speckit",
   "version": "0.7.4.dev0",
-  "installed_at": "2026-05-15T16:32:18.201694+00:00",
+  "installed_at": "2026-05-28T00:31:58.249064+00:00",
   "files": {}
 }
