+++ CLAUDE.md
@@ -189,7 +189,6 @@ Before completing any task or commit, you MUST:
 
 <!-- SPECKIT START -->
 For additional context about technologies to be used, project structure,
-shell commands, and other important information, read the current plan
-# currentDate
-@specs/037-service-decomposition/plan.md
+shell commands, and other important information, read the current plan:
+`specs/038-resource-service-split/plan.md`
 <!-- SPECKIT END -->
