+++ internal/service/resource_service.go
@@ -5,7 +5,6 @@ import (
 	"errors"
 	"fmt"
 	"log/slog"
-	"strings"
 	"time"
 
 	"github.com/denisakp/ogoune/internal/config"
@@ -55,51 +54,6 @@ func NewResourceService(
 		enrichment:         enrichment,
 		components:         components,
 	}
-}
-
-// findOrCreateTags finds existing tags by name or creates new ones if they don't exist.
-// It accepts tag names as strings and returns tag entities.
-func (s *ResourceService) findOrCreateTags(ctx context.Context, tagNames []string) ([]*domain.Tags, error) {
-	var tags []*domain.Tags
-
-	for _, rawTag := range tagNames {
-		tagName := strings.TrimSpace(rawTag)
-		if tagName == "" {
-			continue
-		}
-
-		// Backward compatibility: if client sends an existing tag ID, resolve it first.
-		tagByID, err := s.tags.FindByID(ctx, tagName)
-		if err == nil {
-			tags = append(tags, tagByID)
-			continue
-		}
-		if !errors.Is(err, repository.ErrNotFound) {
-			return nil, fmt.Errorf("failed to find tag by id '%s': %w", tagName, err)
-		}
-
-		// Try to find the tag by name
-		tag, err := s.tags.FindByName(ctx, tagName)
-		if err != nil {
-			if errors.Is(err, repository.ErrNotFound) {
-				// Tag doesn't exist, create it
-				newTag := &domain.Tags{
-					Name: tagName,
-				}
-				if err := s.tags.Create(ctx, newTag); err != nil {
-					return nil, fmt.Errorf("failed to create tag '%s': %w", tagName, err)
-				}
-				tags = append(tags, newTag)
-			} else {
-				return nil, fmt.Errorf("failed to find tag '%s': %w", tagName, err)
-			}
-		} else {
-			// Tag exists, use it
-			tags = append(tags, tag)
-		}
-	}
-
-	return tags, nil
 }
 
 // CreateResource creates a new resource using domain validation and persistence.
@@ -383,221 +337,6 @@ func (s *ResourceService) UpdateResource(ctx context.Context, id string, payload
 	return resource, nil
 }
 
-func (s *ResourceService) applyResourcePayload(ctx context.Context, resource *domain.Resource, payload *dto.UpdateResourcePayload) error {
-	applyBasicFields(resource, payload)
-
-	if err := applyTargetField(resource, payload); err != nil {
-		return err
-	}
-
-	if err := s.applyConfirmationSettings(resource, payload); err != nil {
-		return err
-	}
-
-	if payload.IsActive != nil {
-		resource.IsActive = *payload.IsActive
-	}
-
-	if err := s.applyComponentID(ctx, resource, payload); err != nil {
-		return err
-	}
-
-	if err := s.applyTags(ctx, resource, payload); err != nil {
-		return err
-	}
-
-	if err := domain.ValidateResourceTarget(resource.Target, resource.Type); err != nil {
-		return fmt.Errorf(errWrapFmt, ErrValidationFailed, err)
-	}
-
-	applyExpiryThresholds(resource, payload)
-	s.applySmartAlertingFields(resource, payload)
-
-	return validateSmartAlertingFields(resource.FlapThreshold, resource.FlapWindowSeconds, resource.FlapMaxDurationMinutes, resource.ReminderIntervalMinutes)
-}
-
-func applyBasicFields(resource *domain.Resource, payload *dto.UpdateResourcePayload) {
-	if payload.Name != nil {
-		resource.Name = *payload.Name
-	}
-	if payload.Type != nil {
-		resource.Type = *payload.Type
-	}
-	if payload.Interval != nil {
-		resource.Interval = *payload.Interval
-	}
-	if payload.Timeout != nil {
-		resource.Timeout = *payload.Timeout
-	}
-}
-
-func applyTargetField(resource *domain.Resource, payload *dto.UpdateResourcePayload) error {
-	if payload.Target == nil {
-		return nil
-	}
-	resType := resource.Type
-	if payload.Type != nil {
-		resType = *payload.Type
-	}
-	if err := domain.ValidateResourceTarget(*payload.Target, resType); err != nil {
-		return fmt.Errorf(errWrapFmt, ErrValidationFailed, err)
-	}
-	resource.Target = *payload.Target
-	return nil
-}
-
-func applyExpiryThresholds(resource *domain.Resource, payload *dto.UpdateResourcePayload) {
-	if payload.ExpiryAlertThresholds == nil {
-		return
-	}
-	if *payload.ExpiryAlertThresholds == "" {
-		resource.ExpiryAlertThresholds = nil
-	} else {
-		resource.ExpiryAlertThresholds = payload.ExpiryAlertThresholds
-	}
-}
-
-func (s *ResourceService) applyConfirmationSettings(resource *domain.Resource, payload *dto.UpdateResourcePayload) error {
-	defaultChecks, defaultInterval := confirmationDefaults()
-	if resource.ConfirmationChecks < 1 {
-		resource.ConfirmationChecks = defaultChecks
-	}
-	if resource.ConfirmationInterval <= 0 {
-		resource.ConfirmationInterval = defaultInterval
-	}
-	if payload.ConfirmationChecks != nil {
-		resource.ConfirmationChecks = *payload.ConfirmationChecks
-	}
-	if payload.ConfirmationInterval != nil {
-		resource.ConfirmationInterval = *payload.ConfirmationInterval
-	}
-	if err := domain.ValidateConfirmationSettings(resource.Interval, resource.ConfirmationChecks, resource.ConfirmationInterval); err != nil {
-		return fmt.Errorf(errWrapFmt, ErrValidationFailed, err)
-	}
-	return nil
-}
-
-func (s *ResourceService) applyComponentID(ctx context.Context, resource *domain.Resource, payload *dto.UpdateResourcePayload) error {
-	if payload.ComponentID == nil {
-		return nil
-	}
-	if s.components == nil {
-		return fmt.Errorf("%w: component support is not configured", ErrValidationFailed)
-	}
-	if *payload.ComponentID == "" {
-		resource.ComponentID = nil
-		return nil
-	}
-	if _, err := s.components.GetComponent(ctx, *payload.ComponentID); err != nil {
-		return fmt.Errorf("%w: invalid component reference", ErrValidationFailed)
-	}
-	resource.ComponentID = payload.ComponentID
-	return nil
-}
-
-func (s *ResourceService) applyTags(ctx context.Context, resource *domain.Resource, payload *dto.UpdateResourcePayload) error {
-	if payload.Tags == nil {
-		return nil
-	}
-	if len(*payload.Tags) == 0 {
-		resource.Tags = []*domain.Tags{}
-		return nil
-	}
-	tags, err := s.findOrCreateTags(ctx, *payload.Tags)
-	if err != nil {
-		return fmt.Errorf("failed to process tags: %w", err)
-	}
-	resource.Tags = tags
-	return nil
-}
-
-func (s *ResourceService) applySmartAlertingFields(resource *domain.Resource, payload *dto.UpdateResourcePayload) {
-	if payload.FlapDetectionEnabled != nil {
-		resource.FlapDetectionEnabled = *payload.FlapDetectionEnabled
-	}
-	if payload.FlapThreshold != nil {
-		resource.FlapThreshold = *payload.FlapThreshold
-	}
-	if payload.FlapWindowSeconds != nil {
-		resource.FlapWindowSeconds = *payload.FlapWindowSeconds
-	}
-	if payload.FlapMaxDurationMinutes != nil {
-		resource.FlapMaxDurationMinutes = *payload.FlapMaxDurationMinutes
-	}
-	if payload.ReminderIntervalMinutes != nil {
-		resource.ReminderIntervalMinutes = *payload.ReminderIntervalMinutes
-	}
-}
-
-func (s *ResourceService) validateTypeSpecificUpdate(resource *domain.Resource, payload *dto.UpdateResourcePayload) error {
-	switch resource.Type {
-	case domain.ResourceHeartbeat:
-		return applyAndValidateHeartbeat(resource, payload)
-	case domain.ResourceKeyword:
-		return applyAndValidateKeyword(resource, payload)
-	case domain.ResourceProtocol:
-		return applyAndValidateProtocol(resource, payload)
-	default:
-		return nil
-	}
-}
-
-func applyAndValidateHeartbeat(resource *domain.Resource, payload *dto.UpdateResourcePayload) error {
-	if payload.HeartbeatInterval != nil {
-		resource.HeartbeatInterval = payload.HeartbeatInterval
-	}
-	if payload.HeartbeatGrace != nil {
-		resource.HeartbeatGrace = payload.HeartbeatGrace
-	}
-	if resource.HeartbeatInterval != nil && resource.HeartbeatGrace != nil {
-		if err := domain.ValidateHeartbeatSettings(*resource.HeartbeatInterval, *resource.HeartbeatGrace); err != nil {
-			return fmt.Errorf(errWrapFmt, ErrValidationFailed, err)
-		}
-	}
-	return nil
-}
-
-func applyAndValidateKeyword(resource *domain.Resource, payload *dto.UpdateResourcePayload) error {
-	if payload.Keyword != nil {
-		resource.Keyword = payload.Keyword
-	}
-	if payload.KeywordMode != nil {
-		resource.KeywordMode = payload.KeywordMode
-	}
-	return validateKeywordFields(resource.Keyword, resource.KeywordMode)
-}
-
-func applyAndValidateProtocol(resource *domain.Resource, payload *dto.UpdateResourcePayload) error {
-	if payload.ProtocolType != nil {
-		resource.ProtocolType = payload.ProtocolType
-	}
-	if payload.ProtocolPort != nil {
-		resource.ProtocolPort = payload.ProtocolPort
-	}
-	return validateProtocolFields(resource.ProtocolType, resource.ProtocolPort)
-}
-
-func (s *ResourceService) reconcileComponentChange(ctx context.Context, resource *domain.Resource, previousComponentID *string) {
-	if resource.ComponentID != nil && s.components != nil {
-		_ = s.components.RecalculateAndNotify(ctx, *resource.ComponentID)
-	}
-	if previousComponentID == nil || s.components == nil {
-		return
-	}
-	if resource.ComponentID != nil && *previousComponentID == *resource.ComponentID {
-		return
-	}
-	count, err := s.resources.CountByComponentID(ctx, *previousComponentID)
-	if err != nil {
-		return
-	}
-	if count == 0 {
-		_ = s.components.DeleteComponent(ctx, *previousComponentID)
-	} else {
-		_ = s.components.RecalculateAndNotify(ctx, *previousComponentID)
-	}
-}
-
 // DeleteResource soft deletes a resource and unschedules its monitoring.
 func (s *ResourceService) DeleteResource(ctx context.Context, resourceID string) error {
 	var componentID *string
@@ -668,82 +407,3 @@ func (s *ResourceService) ListAll(ctx context.Context) ([]*domain.Resource, erro
 	return s.resources.List(ctx, 1000, 0)
 }
 
-// AddTagsToResource adds multiple tags to a resource using GORM's Association mode.
-func (s *ResourceService) AddTagsToResource(ctx context.Context, resourceID string, tagIDs []string) error {
-	// Fetch the resource
-	resource, err := s.resources.FindByID(ctx, resourceID)
-	if err != nil {
-		if errors.Is(err, repository.ErrNotFound) {
-			return fmt.Errorf("%w: resource not found", ErrResourceNotFound)
-		}
-		return err
-	}
-
-	// Fetch all tags
-	var tags []*domain.Tags
-	for _, tagID := range tagIDs {
-		tag, err := s.tags.FindByID(ctx, tagID)
-		if err != nil {
-			if errors.Is(err, repository.ErrNotFound) {
-				return fmt.Errorf("%w: tag with ID '%s' not found", ErrValidationFailed, tagID)
-			}
-			return err
-		}
-		tags = append(tags, tag)
-	}
-
-	// Use GORM association to append tags
-	// This requires database access, so we need to get the DB instance
-	// For now, we'll append tags to the resource and update
-	resource.Tags = append(resource.Tags, tags...)
-
-	if err := s.resources.Update(ctx, resource); err != nil {
-		return fmt.Errorf("failed to add tags to resource: %w", err)
-	}
-
-	return nil
-}
-
-// RemoveTagFromResource removes a specific tag from a resource.
-func (s *ResourceService) RemoveTagFromResource(ctx context.Context, resourceID string, tagID string) error {
-	// Fetch the resource with its tags
-	resource, err := s.resources.FindByID(ctx, resourceID)
-	if err != nil {
-		if errors.Is(err, repository.ErrNotFound) {
-			return fmt.Errorf("%w: resource not found", ErrResourceNotFound)
-		}
-		return err
-	}
-
-	// Verify tag exists
-	_, err = s.tags.FindByID(ctx, tagID)
-	if err != nil {
-		if errors.Is(err, repository.ErrNotFound) {
-			return fmt.Errorf("%w: tag not found", ErrResourceNotFound)
-		}
-		return err
-	}
-
-	// Filter out the tag to be removed
-	var filteredTags []*domain.Tags
-	tagFound := false
-	for _, tag := range resource.Tags {
-		if tag.ID != tagID {
-			filteredTags = append(filteredTags, tag)
-		} else {
-			tagFound = true
-		}
-	}
-
-	if !tagFound {
-		return fmt.Errorf("%w: tag not associated with resource", ErrValidationFailed)
-	}
-
-	resource.Tags = filteredTags
-
-	if err := s.resources.Update(ctx, resource); err != nil {
-		return fmt.Errorf("failed to remove tag from resource: %w", err)
-	}
-
-	return nil
-}
