+++ internal/service/resource_tags.go
@@ -1,0 +1,136 @@
+package service
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"strings"
+
+	"github.com/denisakp/ogoune/internal/domain"
+	"github.com/denisakp/ogoune/internal/repository"
+)
+
+// findOrCreateTags finds existing tags by name or creates new ones if they don't exist.
+// It accepts tag names as strings and returns tag entities.
+func (s *ResourceService) findOrCreateTags(ctx context.Context, tagNames []string) ([]*domain.Tags, error) {
+	var tags []*domain.Tags
+
+	for _, rawTag := range tagNames {
+		tagName := strings.TrimSpace(rawTag)
+		if tagName == "" {
+			continue
+		}
+
+		// Backward compatibility: if client sends an existing tag ID, resolve it first.
+		tagByID, err := s.tags.FindByID(ctx, tagName)
+		if err == nil {
+			tags = append(tags, tagByID)
+			continue
+		}
+		if !errors.Is(err, repository.ErrNotFound) {
+			return nil, fmt.Errorf("failed to find tag by id '%s': %w", tagName, err)
+		}
+
+		// Try to find the tag by name
+		tag, err := s.tags.FindByName(ctx, tagName)
+		if err != nil {
+			if errors.Is(err, repository.ErrNotFound) {
+				// Tag doesn't exist, create it
+				newTag := &domain.Tags{
+					Name: tagName,
+				}
+				if err := s.tags.Create(ctx, newTag); err != nil {
+					return nil, fmt.Errorf("failed to create tag '%s': %w", tagName, err)
+				}
+				tags = append(tags, newTag)
+			} else {
+				return nil, fmt.Errorf("failed to find tag '%s': %w", tagName, err)
+			}
+		} else {
+			// Tag exists, use it
+			tags = append(tags, tag)
+		}
+	}
+
+	return tags, nil
+}
+
+// AddTagsToResource adds multiple tags to a resource using GORM's Association mode.
+func (s *ResourceService) AddTagsToResource(ctx context.Context, resourceID string, tagIDs []string) error {
+	// Fetch the resource
+	resource, err := s.resources.FindByID(ctx, resourceID)
+	if err != nil {
+		if errors.Is(err, repository.ErrNotFound) {
+			return fmt.Errorf("%w: resource not found", ErrResourceNotFound)
+		}
+		return err
+	}
+
+	// Fetch all tags
+	var tags []*domain.Tags
+	for _, tagID := range tagIDs {
+		tag, err := s.tags.FindByID(ctx, tagID)
+		if err != nil {
+			if errors.Is(err, repository.ErrNotFound) {
+				return fmt.Errorf("%w: tag with ID '%s' not found", ErrValidationFailed, tagID)
+			}
+			return err
+		}
+		tags = append(tags, tag)
+	}
+
+	// Use GORM association to append tags
+	// This requires database access, so we need to get the DB instance
+	// For now, we'll append tags to the resource and update
+	resource.Tags = append(resource.Tags, tags...)
+
+	if err := s.resources.Update(ctx, resource); err != nil {
+		return fmt.Errorf("failed to add tags to resource: %w", err)
+	}
+
+	return nil
+}
+
+// RemoveTagFromResource removes a specific tag from a resource.
+func (s *ResourceService) RemoveTagFromResource(ctx context.Context, resourceID string, tagID string) error {
+	// Fetch the resource with its tags
+	resource, err := s.resources.FindByID(ctx, resourceID)
+	if err != nil {
+		if errors.Is(err, repository.ErrNotFound) {
+			return fmt.Errorf("%w: resource not found", ErrResourceNotFound)
+		}
+		return err
+	}
+
+	// Verify tag exists
+	_, err = s.tags.FindByID(ctx, tagID)
+	if err != nil {
+		if errors.Is(err, repository.ErrNotFound) {
+			return fmt.Errorf("%w: tag not found", ErrResourceNotFound)
+		}
+		return err
+	}
+
+	// Filter out the tag to be removed
+	var filteredTags []*domain.Tags
+	tagFound := false
+	for _, tag := range resource.Tags {
+		if tag.ID != tagID {
+			filteredTags = append(filteredTags, tag)
+		} else {
+			tagFound = true
+		}
+	}
+
+	if !tagFound {
+		return fmt.Errorf("%w: tag not associated with resource", ErrValidationFailed)
+	}
+
+	resource.Tags = filteredTags
+
+	if err := s.resources.Update(ctx, resource); err != nil {
+		return fmt.Errorf("failed to remove tag from resource: %w", err)
+	}
+
+	return nil
+}
